Backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic[email]
"pymongo[srv]"
python-dotenv